feat: replaced mongoose db connection setup with mongodb & for document retrieval
