+++ src/components/mongodb/DBConnector.tsx
@@ -1,54 +1,46 @@
 import { config } from "@/../config";
-import mongoose from "mongoose";
-import TinkerforgeSensor from "@/models/Tinkerforgesensor.schema";
 
-/*
- * Creates a connection with MongoDB
- */
-async function StartSetup() {
-  console.log("Starting MongoDB connection...");
-  await mongoose.connect(`${config.krakenConfigsMongodbConnectionString}`);
-  console.log("Connection with MongoDB successfully established.");
-}
+const { MongoClient } = require("mongodb");
+const uri = config.krakenConfigsMongodbConnectionString;
+const client = new MongoClient(uri);
 
-/*
- * Validates the MongoDB connection exists and is working
- */
-async function ValidateMongoDBConnection() {
-  console.log("MongoDB ping test - Starting...");
-  if (mongoose.connection.db === undefined) {
-    throw new Error(
-      "MongoDB connection is undefined. Please check the db connector and the MongoDB connection setup.",
-    );
-  } else {
-    await mongoose.connection.db.admin().command({ ping: 1 });
-    console.log("MongoDB ping test - successful");
-  }
+async function connectToDB() {
+  await client.connect();
 }
 
 /*
  * Queries all documents from TinkerforgeSensor collection
  */
-async function getAllDocuments() {
-  return await TinkerforgeSensor.find({});
+export async function getAllDocuments() {
+  const database = await client.db("sensor_config");
+  const sensors = await database.collection("TinkerforgeSensor");
+  let allDocs = await sensors.find({}).toArray();
+  // Converting to plain values
+  // let sensorData = JSON.stringify(Object.values(allDocs), null, 2);
+  // let sensorData = Object.values(allDocs);
+
+  return allDocs;
 }
 
 export default async function DBConnector() {
   let connectionFailed = false;
   try {
-    await StartSetup();
-    await ValidateMongoDBConnection();
+    await connectToDB();
+    let documents = await getAllDocuments();
+    console.log("\nprinting documents start\n");
+    console.log(documents);
+    console.log("\nprinting documents end\n");
   } catch (error) {
+    // Connection couldn't be established
     connectionFailed = true;
     console.log("MongoDB Connection Setup Error:", error);
     return <span>error2</span>;
+  } finally {
+    // Ensures that the client will close when you finish/error
+    await client.close();
   }
-
-  let documents = await getAllDocuments();
-  console.log("\nprinting documents start\n");
-  console.log(documents);
-  console.log("\nprinting documents end\n");
   if (connectionFailed) {
+    // Display connection error
     return <span>error</span>;
   } else {
     return <span>Connection successfully set up</span>;
